+++ test/EliteChroma.Core.Tests/ColorExtensionsTests.cs
@@ -12,16 +12,5 @@ namespace EliteChroma.Core.Tests
         {
             Assert.Throws<ArgumentNullException>("transform", () => ColorExtensions.Transform(Color.Blue, null));
         }
-
-        [Theory]
-        [InlineData(-1, 0)]
-        [InlineData(0, 0)]
-        [InlineData(1, 255)]
-        [InlineData(2, 255)]
-        public void EnsureBoundKeepValuesWithinRange(double input, byte expected)
-        {
-            var c = ColorExtensions.Combine(Color.Black, Color.White, input);
-            Assert.Equal(expected, c.R);
-        }
     }
 }